Added the ability to use Protege catalog-v00.xml files to map ontology URIs to files (local and remote). See how to use this in the GenerateRefactoringReport.
The example depends on setting the run JVM property "isf.svn.truk.root".
Eclipse can help with this by creating a variable in the workspace
preferences that you can then use consistently in various run
configurations. The run configuration for the example has an entry in
the JVM properties that looks like:
-D${SVN_ROOT_VARIABLE}
and the variable is defined in the preferences like:
name: SVN_ROOT_VARIABLE
value: isf.svn.trunk.root=c:\path\to\local\trunk\root  (the
slash direction might be system dependent)